[refactor] 카프카 설정 디렉토리 리팩토링
